added menu page
All we have to do now for the site is insert the correct links and data for menu and restaurant pages. We also have to go and spell restaurant correctly cuz apparently i cant spell. We can also add responsiveness if we want. And once the database is done we can add the rest of the restaurants.